+++ native/c/zds.cpp
@@ -893,6 +893,7 @@ int zds_read_vsam(ZDS *zds, std::string ddname, std::string &response)
   }
 
   int lines_read = 0;
+  zds->has_more = 0;
   while (true)
   {
     rc = ZDSRIVSM(zds, ioc);
@@ -901,6 +902,11 @@ int zds_read_vsam(ZDS *zds, std::string ddname, std::string &response)
       response.append(ioc->buffer, ioc->ifgacb.acblrecl);
       response.append(1, '\n');
       lines_read++;
+      if (lines_read >= zds->max_lines)
+      {
+        zds->has_more = 1;
+        break;
+      }
     }
     else if (rc == RTNCD_WARNING)
     {
@@ -911,19 +917,11 @@ int zds_read_vsam(ZDS *zds, std::string ddname, std::string &response)
       ZDSCIVSM(zds, ioc);
       return rc;
     }
-    if (lines_read >= zds->max_lines)
-    {
-      break;
-    }
   }
 
-  rc = ZDSCIVSM(zds, ioc);
-  if (rc != RTNCD_SUCCESS)
-  {
-    return rc;
-  }
+  zds->returned_lines = lines_read;
 
-  return RTNCD_SUCCESS;
+  return ZDSCIVSM(zds, ioc);
 }
 
 /**